Use module util HexDump in StructTest
The hidden API HexDump causes issues, as coverage tests jarjar its
package even though it is a dependency on a hidden platform class, and
not included in test packages.

Use the HexDump utility in the com.android.net.module.util package
instead (no import is necessary as this is the same package).

Bug: 187935317
Test: atest StructTest
Change-Id: If319a3c20a750c890f8400f1a7104462b41ed556