Add wrappers for the latest stable version of netd_aidl_interface-V*-cpp
Currently, we need to update linked AILD interface in several places
when we freeze a new interface version. Otherwise, it will have multiple
versions of an AIDL lib and waste space/memory (b/220392885).

To avoid manually update all AILD interface users every time, adds
wrappers for the latest version of netd_aidl_interface-V*-cpp.

The reason to use cc_default as a wrapper instead of cc_library is that
it doesn't create an extra .so.

Bug: 220392885
Test: TH
Change-Id: I281b2e5f7d04fadfae0ede4db7565e9f222dbe95